feat(rsu): moved RsuOnlineStatusController and corresponding tests
- Updated RSU online status to use the following endpoint: /devices/rsus/online-status
- Updated webapp API slice to reflect the new endpoint structure.
